requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
altgraph==0.17
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
daphne==2.5.0
defusedxml==0.7.0rc1
Django==3.1.4
django-bootstrap-static==4.2.1
django-cors-headers==3.5.0
django-cropperjs==0.0.5
django-filter==2.4.0
django-grappelli==2.14.2
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
dominate==2.6.0
freeze==3.0
future==0.18.2
geojson==2.5.0
gitdb==4.0.5
GitPython==3.1.11
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
multidict==4.7.6
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.40
optional-django==0.1.0
packaging==20.4
pefile==2019.4.18
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyInstaller==3.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyowm==3.0.0
pyparsing==2.4.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.2
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
smmap==3.0.4
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
yarl==1.4.2
zipp==3.1.0
zope.interface==5.1.0
